Make support for multiple codecs to really work.
That was brain dead. To put a 64 bit wide flag value (well, with just the lower 15 bits affecting us, but even so!) inside a 7 bit wide field with no transform and expecting it to work? Crazy!
This is the generic version, which can be optimized taking into account we set up the encoder per channel created / answered.